# club_cms/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Club CMS API",
        "description": "Documentación de la API para el Club CMS",
        "contact": {
            "email": "soporte@clubcms.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Base URL para todos los endpoints"
        }
    ],
    "paths": {
...
        "/reservations": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Listar reservas",
                "description": "Obtiene una lista paginada de todas las reservas. Accesible para usuarios y administradores autenticados.",
                "operationId": "c0ead7ee8da750a341eeff509f3af8b9",
                "responses": {
                    "200": {
                        "description": "Listado de reservas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "reservations": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ReservationResource"
                                            }
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_pages": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "total_reservations": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay reservas disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener las reservas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Crear una nueva reserva",
                "description": "Crea una nueva reserva en el sistema. Valida límites de reservas (máximo 3 por día) y previene duplicados.",
                "operationId": "ab4d7952bb3ef1ccd7724e8d862b034b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reserva creada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationCreateResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Miembro o cancha no encontrados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación - Límites de reserva excedidos o datos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear la reserva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/reservations/{id}": {
            "get": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Mostrar detalles de una reserva",
                "description": "Obtiene los detalles de una reserva específica. Accesible para usuarios y administradores autenticados.",
                "operationId": "49499ada3f4d3680e147782d33573252",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle de la reserva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "reservation": {
                                            "$ref": "#/components/schemas/ReservationResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener la reserva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Actualizar una reserva existente",
                "description": "Actualiza los datos de una reserva existente. Valida límites de reservas y previene duplicados.",
                "operationId": "4a219e5cfd9760d12da084929201d042",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReservationRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reserva actualizada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReservationUpdateResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Reserva no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación - Límites de reserva excedidos o datos inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar la reserva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Reservations"
                ],
                "summary": "Eliminar una reserva",
                "description": "Elimina una reserva del sistema. Accesible para usuarios y administradores autenticados.",
                "operationId": "b97e0df7036fa97413d6b819cf61a11e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la reserva a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reserva eliminada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Reserva eliminada correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al eliminar la reserva",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ReservationRequest": {
                "required": [
                    "member_id",
                    "court_id",
                    "date",
                    "hour"
                ],
                "properties": {
                    "member_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "court_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    },
                    "hour": {
                        "type": "string",
                        "format": "time",
                        "example": "14:00"
                    }
                },
                "type": "object"
            },
            "ReservationResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "member_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "court_name": {
                        "type": "string",
                        "example": "Pista 1"
                    },
                    "date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    },
                    "hour": {
                        "type": "string",
                        "format": "time",
                        "example": "14:00"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "description": "Respuesta estándar para errores internos del servidor",
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Mensaje de error seguro que no expone detalles técnicos",
                        "type": "string",
                        "example": "Error interno del servidor al procesar la solicitud."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "description": "Respuesta cuando falla la validación de los datos de la petición",
                "properties": {
                    "ok": {
                        "description": "Indica que la operación falló",
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Mensaje de error general",
                        "type": "string",
                        "example": "Error de validación"
                    },
                    "errors": {
                        "description": "Listado de errores de validación por campo",
                        "type": "object",
                        "example": {
                            "email": [
                                "El correo electrónico ya está registrado por otro miembro."
                            ],
                            "member_id": [
                                "El miembro ya tiene el máximo de 3 reservas permitidas para esta fecha."
                            ],
                            "court_id": [
                                "Ya existe una reserva para esta cancha en la misma fecha y hora."
                            ],
                            "date": [
                                "La fecha no puede ser anterior a hoy."
                            ],
                            "hour": [
                                "El horario debe estar entre las 08:00 y las 21:00."
                            ],
                            "name": [
                                "El campo nombre es obligatorio."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "ReservationCreateResponse": {
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Reserva creada correctamente"
                    },
                    "reservation": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 25
                            },
                            "member_id": {
                                "type": "integer",
                                "example": 1
                            },
                            "court_name": {
                                "type": "string",
                                "example": "Pista 1"
                            },
                            "date": {
                                "type": "string",
                                "format": "date",
                                "example": "2023-01-01"
                            },
                            "hour": {
                                "type": "string",
                                "format": "time",
                                "example": "14:00"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}